+++ src/SparkRoot.js
@@ -1,10 +1,12 @@
 import React, { useEffect, useState, useCallback } from 'react';
 import Pbf from 'pbf'
 import history from 'history/browser';
+import Homepage from './Homepage';
 import { SamplerData, HeapData } from './proto'
-import { Sampler, MappingsMenu, labelData } from './types/Sampler';
-import { getMappingsInfo, requestMappings } from './mappings'
-import { Heap } from './types/Heap';
+import { Sampler, labelData } from './sampler/Sampler';
+import { MappingsMenu } from './sampler/MappingsMenu';
+import { getMappingsInfo, requestMappings } from './sampler/mappings'
+import { Heap } from './heap/Heap';
 
 import sparkLogo from './assets/spark-logo.svg'
 
@@ -16,53 +18,39 @@ const LOADED_PROFILE_DATA = Symbol();
 const LOADED_HEAP_DATA = Symbol();
 const PAGE_NOT_FOUND = Symbol();
 
-function Header({ isViewer, mappings, setMappings }) {
-    return (
-        <div id="header">
-            <a href="/" id="logo">
-                <img src={sparkLogo} alt="" width="32px" height="32px" />
-                {isViewer ? <h1>spark viewer</h1> : <h1>spark</h1>}
-            </a>
-            {mappings ? <MappingsMenu mappings={mappings} setMappings={setMappings} /> : null}
-        </div>
-    )
-}
+/**
+ * Gets the payload code from the URL.
+ * @returns the code, or undefined
+ */
+function getUrlCode() {
+    const path = window.location.pathname;
+    const hash = window.location.hash;
 
-function Footer() {
-    return (
-        <div id="footer">
-            <a href="https://github.com/lucko/spark">spark</a> and <a href="https://github.com/lucko/spark-viewer">spark-viewer</a> are based on WarmRoast by sk89q.<br />
-            Copyright &copy; 2018-2021 <a href="https://github.com/lucko">lucko</a>, <a href="https://github.com/astei">astei</a> & spark contributors
-        </div>
-    )
+    let code;
+    if (path === '/' && /^#[a-zA-Z0-9]+$/.test(hash)) {
+        code = hash.substring(1);
+        // change URL to remove the hash
+        history.replace({
+            pathname: code,
+            hash: ''
+        });
+    } else if (/^\/[a-zA-Z0-9]+$/.test(path)) {
+        code = path.substring(1);
+    }
+    return code;
 }
 
 export default function SparkRoot() {
-    const [code] = useState(() => {
-        const path = window.location.pathname;
-        const hash = window.location.hash;
-    
-        let code;
-        if (path === '/' && /^#[a-zA-Z0-9]+$/.test(hash)) {
-            code = hash.substring(1);
-            // change URL to remove the hash
-            history.replace({
-                pathname: code,
-                hash: ''
-            });
-        } else if (/^\/[a-zA-Z0-9]+$/.test(path)) {
-            code = path.substring(1);
-        }
-        return code;
-    });
-
+    const [code] = useState(getUrlCode);
     const [status, setStatus] = useState(code ? LOADING_DATA : window.location.pathname === '/' ? HOMEPAGE : PAGE_NOT_FOUND);
     const [loaded, setLoaded] = useState(null);
     const [mappingsInfo, setMappingsInfo] = useState(null);
     const [mappings, setMappings] = useState({func: _ => {}});
     const [mappingsType, setMappingsType] = useState(null);
 
-    const onMappingsRequest = useCallback((type) => {
+    // Function called whenever the user picks mappings, either
+    // from the input dropdown, or 'auto' when mappings info is loaded.
+    const onMappingsRequest = useCallback(type => {
         if (mappingsType !== type) {
             setMappingsType(type);
             requestMappings(type, mappingsInfo, loaded).then(func => {
@@ -71,15 +59,36 @@ export default function SparkRoot() {
         }
     }, [mappingsType, mappingsInfo, loaded]);
 
+    // Wait for mappingsInfo and data ('loaded') to be populated,
+    // then run a mappings request for 'auto'.
     useEffect(() => {
         if (!mappingsType && mappingsInfo && loaded) {
             onMappingsRequest('auto');
         }
     }, [mappingsType, mappingsInfo, loaded, onMappingsRequest]);
 
+    // On page load, if status is set to LOADING_DATA, make
+    // a request to bytebin to load the payload
     useEffect(() => {
         if (status !== LOADING_DATA) {
             return;
+        }
+
+        function initSampler(buf) {
+            setStatus(PARSING_DATA);
+            const pbf = new Pbf(new Uint8Array(buf));
+            const data = SamplerData.read(pbf);
+            labelData(data.threads, 0);
+            setLoaded(data);
+            setStatus(LOADED_PROFILE_DATA);
+        }
+
+        function initHeap(buf) {
+            setStatus(PARSING_DATA);
+            const pbf = new Pbf(new Uint8Array(buf));
+            const data = HeapData.read(pbf);
+            setLoaded(data);
+            setStatus(LOADED_HEAP_DATA);
         }
 
         async function onLoad() {
@@ -94,23 +103,11 @@ export default function SparkRoot() {
                 if (type === 'application/x-spark-sampler') {
                     // request mappings metadata in the background
                     getMappingsInfo().then(setMappingsInfo);
-
                     const buf = await req.arrayBuffer();
-                    setStatus(PARSING_DATA);
-                    const pbf = new Pbf(new Uint8Array(buf));
-                    const data = SamplerData.read(pbf);
-                    labelData(data);
-
-                    setLoaded(data);
-                    setStatus(LOADED_PROFILE_DATA);
+                    initSampler(buf);
                 } else if (type === 'application/x-spark-heap') {
                     const buf = await req.arrayBuffer();
-                    setStatus(PARSING_DATA);
-                    const pbf = new Pbf(new Uint8Array(buf));
-                    const data = HeapData.read(pbf);
-
-                    setLoaded(data);
-                    setStatus(LOADED_HEAP_DATA);
+                    initHeap(buf);
                 } else {
                     setStatus(FAILED_DATA);
                 }
@@ -126,52 +123,61 @@ export default function SparkRoot() {
     let contents;
     switch (status) {
         case HOMEPAGE:
-            contents = (
-                <div id="intro">
-                    <h1># spark</h1>
-                    <p>spark is a performance profiling plugin/mod for Minecraft clients, servers and proxies.</p>
-                    <p>spark is made up of three separate components:</p>
-                    <ul>
-                        <li><b>CPU Profiler</b>: Diagnose performance issues with your server.</li>
-                        <li><b>Memory Inspection</b>: Diagnose memory issues with your server.</li>
-                        <li><b>Server Health Reporting</b>: Keep track of your servers overall health.</li>
-                    </ul>
-                    <p>You can find <a href="https://spark.lucko.me/docs">documentation</a> for spark on our docs website.</p>
-                    <p>You can <a href="https://ci.lucko.me/job/spark/">download</a> the latest version of the plugin from Jenkins.</p>
-                    <p>The source code and more information about the spark plugin is available on <a href="https://github.com/lucko/spark">GitHub</a>.</p>
-
-                    <h2># Viewer</h2>
-                    <p>This website contains an online viewer for spark profiles.</p>
-                    <p>It is written using React, and open-source'd on GitHub. Any contributions are most welcome!</p>
-                    <p>Uploaded content is stored centrally and retained for 60 days.</p>
-                </div>
-            )
+            contents = <Homepage />
             break
         case PAGE_NOT_FOUND:
-            contents = <div className="banner-notice">404 - Page Not Found</div>
+            contents = <BannerNotice>404 - Page Not Found</BannerNotice>
             break
         case LOADING_DATA:
-            contents = <div className="banner-notice">Downloading...</div>
+            contents = <BannerNotice>Downloading...</BannerNotice>
             break
         case PARSING_DATA:
-            contents = <div className="banner-notice">Rendering...</div>
+            contents = <BannerNotice>Rendering...</BannerNotice>
             break
         case FAILED_DATA:
-            contents = <div className="banner-notice">Unable to load the data. Perhaps it expired? Are you using a recent version?</div>
+            contents = <BannerNotice>Unable to load the data. Perhaps it expired? Are you using a recent version?</BannerNotice>
             break
         case LOADED_PROFILE_DATA:
-            contents = <Sampler data={loaded} mappings={mappings}/>
+            contents = <Sampler data={loaded} mappings={mappings} />
             break
         case LOADED_HEAP_DATA:
-            contents = <Heap data={loaded}/>
+            contents = <Heap data={loaded} />
             break
         default:
-            contents = <div className="banner-notice">Unknown state - this is a bug.</div>
+            contents = <BannerNotice>Unknown state - this is a bug.</BannerNotice>
             break
     }
     return <>
-        <Header isViewer={code} mappings={mappingsInfo} setMappings={onMappingsRequest} />
+        <Header isViewer={!!code} mappings={mappingsInfo} setMappings={onMappingsRequest} />
         {contents}
         <Footer />
     </>
+}
+
+function Header({ isViewer, mappings, setMappings }) {
+    return (
+        <div id="header">
+            <a href="/" id="logo">
+                <img src={sparkLogo} alt="" width="32px" height="32px" />
+                {isViewer
+                    ? <h1>spark viewer</h1>
+                    : <h1>spark</h1>
+                }
+            </a>
+            {!!mappings && <MappingsMenu {...{mappings, setMappings}} />}
+        </div>
+    )
+}
+
+function Footer() {
+    return (
+        <div id="footer">
+            <a href="https://github.com/lucko/spark">spark</a> and <a href="https://github.com/lucko/spark-viewer">spark-viewer</a> are based on WarmRoast by sk89q.<br />
+            Copyright &copy; 2018-2021 <a href="https://github.com/lucko">lucko</a>, <a href="https://github.com/astei">astei</a> & spark contributors
+        </div>
+    )
+}
+
+function BannerNotice(props) {
+    return <div className="banner-notice">{props.children}</div>
 }
